Install typefaces on new hosts
This install typefaces that I use in my code editor. I use Fira Code for the actual code documents, and Geist Mono for the editor interface. Before installing any typefaces, the script will assert that the operating system is supported to assign the correct installation directory, and abort the script if the operating system isn't supported or is unknown.

The script will take files from the repository (`./assets/typefaces/*`) and copy them over to where typefaces are installed on the system. It'll avoid copying files that already exist. During runtime, the script will report how many files are found in the repository, and explicitly account for the action taken on each one (installed or skipped). It'll also sum up how many of the files were installed. Additionally, the script will give warnings if the repository doesn't have the expected directory, or if the directory is empty.

Two utility scripts are included; one for assigning the installation directory for typefaces depending on the operating system, one for assisting with colour assignment in the terminal.

Finally, ignore Apple .DS_Store files. They are files used by MacOS to keep track of location (X & Y coordinates) of files and folders in the file explorer and are of no interest to this repository.