Refine documentation, and clarify versioning scheme
To make this a turn-key solution, include a command that'll encompass more of what is required, including pulling the repository and changing directory. The intent is to lower the required knowledge to publish these dotfiles from scratch on a new host and to make the process, albeit arguably negligibly so, faster (less commands to type), easier (less commands to remember), and more comfortable (the computer does the work instead of you).
Note that the home directory may contain spaces, so it's wrapped up in quotation marks.

Additionally, elaborate on versioning practices through `Versioning.md`. This will clarify the rationale behind using two different versioning schemes and how I differentiate [feature states](../docs/Changelog.md#features-states).