Make typeface variables contextually relevant
Moving the assignment of `FONT_INSTALLATION_DIRECTORY` to the `typefaces.sh` make the decision and the variable more contextually bound to each other. After this change, `system.sh` only determines the operating system, and whatever is contextually relevant for typefaces is handled in `typefaces.sh`.

Use explicit variable syntax in installation command, to make it even more abundantly clear that they're variables.

Finally, stylise the header in `README.md`, just because.