Fewer options to scan while navigating `README.md`
Assumption is that Utilities isn't really interesting, or at least less interesting than other features, as they merely help with the repository whilst not providing much in terms of features to the computer itself, so I down-play it to be a sub-section of the Features topic. This way, there is less for a user to scan through while trying to navigate the repository.